students/api_views.py
from rest_framework import generics
from .models import Grade, Subject
from .serializers import GradeSerializer, SubjectSerializer

class GradeListCreateAPIView(generics.ListCreateAPIView):
    queryset = Grade.objects.all()
    serializer_class = GradeSerializer

class SubjectListCreateAPIView(generics.ListCreateAPIView):
    queryset = Subject.objects.all()
    serializer_class = SubjectSerializer